Switch to doctrine-migrations and symfony/console
This should fix all issues with CodeIgniter bogus CLI detection, which
shows HTML on console depending on your environment.

Migrations are now based on doctrine-migrations, so more DB drivers should
be supported now.

This commit adds required migrations to upgrade from 1.2.6.2, and migrations
required for a fresh installation.

Docs have to be updated.