Improve the ideas behind Event/EventInstances
Event instances had to contain the event URL and Etag to make them suitable for
the frontend. I didn't like this approach, as they were not having a Single
Responsibility.

I have added a new FullCalendarEvent class that stores information about a
CalendarObject and each event instance that it contains for a range of dates.