Create a controller for each operation
Not actually a controller, because of CodeIgniter limitations, but every
controller is now on its own class, and should be really easy to switch to another
framework without touching much code.

There is now a 'JSONController' base controller which manages all exceptions
and provides a basic skeleton for every other controller.