Fix access to files when deployed in subdirectory
When deployed to a subdir, the app.request.basepath variable
does not contain an ending slash.
This prevents the favico to load.
Fix it by preprending a slash.

Also fix a similar issue when loading es5-shim.js

The other assets were not affected because they already got
a prepending slash.